wip: UI for grid calculations and time step selection

+++ ApplicationLibCode/ProjectDataModel/RimEclipseStatisticsCaseEvaluator.h
@@ -54,7 +54,7 @@ class RimEclipseStatisticsCaseEvaluator
 {
 public:
     RimEclipseStatisticsCaseEvaluator( const std::vector<RimEclipseCase*>& sourceCases,
-                                       const std::vector<size_t>&          timeStepIndices,
+                                       const std::vector<int>&             timeStepIndices,
                                        const RimStatisticsConfig&          statisticsConfig,
                                        RigEclipseCaseData*                 destinationCase,
                                        RimIdenticalGridCaseGroup*          identicalGridCaseGroup );
@@ -105,7 +105,7 @@ private:
 
 private:
     std::vector<RimEclipseCase*> m_sourceCases;
-    std::vector<size_t>          m_timeStepIndices;
+    std::vector<int>             m_timeStepIndices;
 
     size_t                     m_reservoirCellCount;
     RimStatisticsConfig        m_statisticsConfig;
